Cleaned up, made GardenVexor buildable. Updated SDL libraries and added them.
Most commits have just been to put the code somewhere, but it's nice to leave it in a compilable state.
No SDL .lib files existed because the .gitignore removed them.
I was apparently in the middle of changing the way engine.cpp stored gameobjects from a map to a vector.
All build paths in visual studio were hard-coded as well.
Removed x86 as a build option, It's not like anyone else will build this.