Corrected the concept of rotation
Rotation in gameobjects was stored as a vector3d, which is fine, but confusing, since in 90% of all cases all you care about is the z rotation (axis into the screen).
Now all rotation functions make more sense.
If I want to use the other rotational dimensions in the future I have to deal with it then.